+++ shvatka/core/waiver/interactors.py
@@ -2,7 +2,7 @@ from shvatka.core.waiver.adapters import WaiverVoteAdder, WaiverVoteGetter
 from shvatka.core.interfaces.identity import IdentityProvider
 from shvatka.core.models import dto
 from shvatka.core.models.enums import Played
-from shvatka.core.services.waiver import check_allow_approve_waivers, get_vote_to_voted, add_vote
+from shvatka.core.waiver.services import check_allow_approve_waivers, get_vote_to_voted, add_vote
 from shvatka.core.utils import exceptions
