+++ shvatka/core/waiver/services.py
