+++ shvatka/tgbot/views/waiver.py
@@ -5,7 +5,7 @@ from aiogram.utils.text_decorations import html_decoration as hd
 
 from shvatka.core.models import dto
 from shvatka.core.models.enums.played import Played
-from shvatka.core.services.waiver import get_vote_to_voted
+from shvatka.core.waiver.services import get_vote_to_voted
 from shvatka.core.views.texts import WAIVER_STATUS_MEANING
 from shvatka.infrastructure.db.dao.complex2.waiver import WaiverVoteGetterImpl
 from shvatka.infrastructure.db.dao.holder import HolderDao
